+++ examples/svelte/src/routes/examples/edges/ButtonEdge.svelte
@@ -4,27 +4,18 @@
 	type $$Props = EdgeProps;
 
 	export let id: $$Props['id'] = '';
-	export let source: $$Props['source'] = '';
-	export let target: $$Props['target'] = '';
-	export let animated: $$Props['animated'] = undefined;
-	export let selected: $$Props['selected'] = undefined;
-	export let label: $$Props['label'] = undefined;
-	export let labelStyle: $$Props['labelStyle'] = undefined;
-	export let data: $$Props['data'] = undefined;
 	export let style: $$Props['style'] = undefined;
-	export let markerStart: $$Props['markerStart'] = undefined;
 	export let markerEnd: $$Props['markerEnd'] = undefined;
-	export let interactionWidth: $$Props['interactionWidth'] = undefined;
 
 	export let sourceX: $$Props['sourceX'];
 	export let sourceY: $$Props['sourceY'];
 	export let sourcePosition: $$Props['sourcePosition'];
-	export let sourceHandleId: $$Props['sourceHandleId'] = undefined;
 
 	export let targetX: $$Props['targetX'];
 	export let targetY: $$Props['targetY'];
 	export let targetPosition: $$Props['targetPosition'];
-	export let targetHandleId: $$Props['targetHandleId'] = undefined;
+
+	$$restProps;
 
 	$: [edgePath, labelX, labelY] = getBezierPath({
 		sourceX,
@@ -34,18 +25,6 @@
 		targetY,
 		targetPosition
 	});
-
-	source;
-	target;
-	animated;
-	selected;
-	data;
-	label;
-	labelStyle;
-	markerStart;
-	interactionWidth;
-	sourceHandleId;
-	targetHandleId;
 </script>
 
 <BaseEdge path={edgePath} {markerEnd} {style} />