+++ examples/svelte/src/routes/examples/stress/+page.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import { writable } from 'svelte/store';
 	import {
 		SvelteFlow,
 		Controls,
@@ -12,8 +11,8 @@
 
 	import '@xyflow/svelte/dist/style.css';
 
-	const yNodes = 25;
-	const xNodes = 25;
+	const yNodes = 50;
+	const xNodes = 50;
 
 	const nodeItems: Node[] = [];
 	const edgeItems: Edge[] = [];