+++ examples/svelte/src/routes/examples/subflows/+page.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import { writable } from 'svelte/store';
 	import {
 		SvelteFlow,
 		Controls,
@@ -18,7 +17,7 @@
 		default: DebugNode
 	};
 
-	const nodes = writable<Node[]>([
+	let nodes = $state<Node[]>([
 		{
 			id: '1',
 			type: 'input',
@@ -95,7 +94,7 @@
 		}
 	]);
 
-	const edges = writable([
+	let edges = $state([
 		{
 			id: 'e1-2',
 			source: '1',
@@ -115,13 +114,9 @@
 		{ id: 'e4a-4b2', source: '4a', target: '4b2', zIndex: 100 },
 		{ id: 'e4b1-4b2', source: '4b1', target: '4b2' }
 	]);
-
-	$: {
-		console.log($nodes);
-	}
 </script>
 
-<SvelteFlow {nodes} {edges} {nodeTypes} fitView minZoom={0.1} maxZoom={2.5}>
+<SvelteFlow bind:nodes bind:edges {nodeTypes} fitView minZoom={0.1} maxZoom={2.5}>
 	<Controls />
 	<Background variant={BackgroundVariant.Dots} />
 	<MiniMap />