+++ packages/svelte/src/lib/container/NodeRenderer/NodeRenderer.svelte
@@ -33,7 +33,7 @@
             updates.set(id, {
               id,
               nodeElement: entry.target as HTMLDivElement,
-              forceUpdate: true
+              force: true
             });
           });