+++ packages/svelte/src/lib/container/SvelteFlow/SvelteFlow.svelte
@@ -191,6 +191,7 @@
   {onconnectend}
   {onbeforedelete}
 />
+
 {#each nodes as node (node.id)}
   <NodeUpdate id={node.id} userNode={node} />
 {/each}