+++ packages/svelte/src/lib/plugins/Minimap/Minimap.svelte
@@ -9,10 +9,12 @@
   import cc from 'classcat';
   import {
     getBoundsOfRects,
+    getInternalNodesBounds,
     getNodeDimensions,
     getNodePositionWithOrigin,
     getNodesBounds,
-    nodeHasDimensions
+    nodeHasDimensions,
+    type Rect
   } from '@xyflow/system';
 
   import { useStore } from '$lib/store';