+++ packages/system/src/xydrag/XYDrag.ts
@@ -119,6 +119,8 @@ export function XYDrag<OnNodeDrag extends (e: any, nodes: any, node: any) => voi
         updateNodePositions,
       } = getStoreItems();
 
+      console.log(dragItems);
+
       lastPos = { x, y };
 
       let hasChange = false;
@@ -229,6 +231,8 @@ export function XYDrag<OnNodeDrag extends (e: any, nodes: any, node: any) => voi
         unselectNodesAndEdges,
       } = getStoreItems();
 
+      console.log('drag started');
+
       dragStarted = true;
 
       if ((!selectNodesOnDrag || !isSelectable) && !multiSelectionActive && nodeId) {
@@ -245,6 +249,8 @@ export function XYDrag<OnNodeDrag extends (e: any, nodes: any, node: any) => voi
       const pointerPos = getPointerPosition(event.sourceEvent, { transform, snapGrid, snapToGrid });
       lastPos = pointerPos;
       dragItems = getDragItems(nodeLookup, nodesDraggable, pointerPos, isNodeSelected, nodeId);
+
+      console.log(dragItems);
 
       if (dragItems.size > 0 && (onDragStart || onNodeDragStart || (!nodeId && onSelectionDragStart))) {
         const [currentNode, currentNodes] = getEventHandlerParams({
@@ -264,6 +270,7 @@ export function XYDrag<OnNodeDrag extends (e: any, nodes: any, node: any) => voi
 
     const d3DragInstance = drag()
       .on('start', (event: UseDragEvent) => {
+        console.log('starty');
         const { domNode, nodeDragThreshold, transform, snapGrid, snapToGrid } = getStoreItems();
 
         if (nodeDragThreshold === 0) {
@@ -276,6 +283,7 @@ export function XYDrag<OnNodeDrag extends (e: any, nodes: any, node: any) => voi
         mousePosition = getEventPosition(event.sourceEvent, containerBounds!);
       })
       .on('drag', (event: UseDragEvent) => {
+        console.log('draggin');
         const { autoPanOnNodeDrag, transform, snapGrid, snapToGrid, nodeDragThreshold } = getStoreItems();
         const pointerPos = getPointerPosition(event.sourceEvent, { transform, snapGrid, snapToGrid });
 
@@ -303,6 +311,7 @@ export function XYDrag<OnNodeDrag extends (e: any, nodes: any, node: any) => voi
         }
       })
       .on('end', (event: UseDragEvent) => {
+        console.log('drag end');
         if (!dragStarted) {
           return;
         }